Move passport to single file
