+++ server/config.js
@@ -1,0 +1,3 @@
+module.export = {
+  isProduction: process.env.NODE_ENV === 'production'
+}
