+++ server/middleware/index.js
@@ -1,3 +1,4 @@
 exports.auth = require('./auth')
 exports.error = require('./error')
 exports.notfound = require('./notfound')
+exports.passport = require('./passport')
