+++ server/routes/index.js
@@ -1,4 +1,4 @@
-let r = (module.exports = require('express').Router())
+const r = (module.exports = require('express').Router())
 
 r.use(require('./account'))
 r.use(require('./auth'))
